split translator.cs, add deepL spoofing
