+++ ChatTranslated/Translate/Translator.cs
@@ -1,23 +1,17 @@
 using ChatTranslated.Utils;
 using Dalamud.Networking.Http;
 using GTranslate.Translators;
-using Newtonsoft.Json.Linq;
 using System;
-using System.IO;
 using System.Net;
 using System.Net.Http;
-using System.Reflection;
-using System.Text;
-using System.Text.Json;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using static ChatTranslated.Configuration;
 
 namespace ChatTranslated.Translate
 {
-    internal class Translator
+    internal static class Translator
     {
-        private static readonly HttpClient HttpClient = new(new SocketsHttpHandler
+        internal static readonly HttpClient HttpClient = new(new SocketsHttpHandler
         {
             ConnectCallback = new HappyEyeballsCallback().ConnectCallback,
             AutomaticDecompression = DecompressionMethods.All
@@ -25,21 +19,11 @@ namespace ChatTranslated.Translate
         {
             DefaultRequestVersion = HttpVersion.Version30,
             DefaultVersionPolicy = HttpVersionPolicy.RequestVersionOrLower,
-
             Timeout = TimeSpan.FromSeconds(10)
         };
-        
+
         public static GoogleTranslator GTranslator = new(HttpClient);
         public static BingTranslator BingTranslator = new(HttpClient);
-
-        private const string DefaultContentType = "application/json";
-        private static readonly string? Cfv2 = ReadSecret("ChatTranslated.Resources.cfv2.secret").Replace("\n", string.Empty);
-
-        private static string ReadSecret(string resourceName)
-        {
-            using var stream = Assembly.GetExecutingAssembly().GetManifestResourceStream(resourceName);
-            return stream == null ? "" : new StreamReader(stream).ReadToEnd();
-        }
 
         public static async Task<string> Translate(string text, string targetLanguage, TranslationMode? translationMode = null)
         {
@@ -50,151 +34,41 @@ namespace ChatTranslated.Translate
 
             return mode switch
             {
-                Configuration.TranslationMode.MachineTranslate => await MachineTranslate(text, targetLanguage),
-                Configuration.TranslationMode.DeepL_API => await DeepLTranslate(text, targetLanguage),
-                Configuration.TranslationMode.OpenAI_API => await OpenAITranslate(text, targetLanguage),
-                Configuration.TranslationMode.LLMProxy => await LLMProxyTranslate(text, targetLanguage),
+                Configuration.TranslationMode.MachineTranslate => await MachineTranslate.Translate(text, targetLanguage),
+                Configuration.TranslationMode.DeepL_API => await DeepLTranslate.Translate(text, targetLanguage),
+                Configuration.TranslationMode.OpenAI_API => await OpenAITranslate.Translate(text, targetLanguage),
+                Configuration.TranslationMode.LLMProxy => await LLMProxyTranslate.Translate(text, targetLanguage),
                 _ => text
             };
         }
 
-        private static async Task<string> MachineTranslate(string text, string targetLanguage)
+        public static async Task<string> DetermineLanguage(string messageText)
         {
+            string? langStr = null;
             try
             {
-                var result = await GTranslator.TranslateAsync(text, targetLanguage);
-                return result.Translation;
-            }
-            catch
-            {
-                try
-                {
-                    var result = await BingTranslator.TranslateAsync(text, targetLanguage);
-                    return result.Translation;
-                }
-                catch
-                {
-                    return text;
-                }
-            }
-        }
-
-        public static async Task<string> DeepLTranslate(string text, string targetLanguage)
-        {
-            if (TryGetLanguageCode(targetLanguage, out var languageCode))
-            {
-                var requestBody = new { text = new[] { text }, target_lang = languageCode, context = "FFXIV, MMORPG" };
-                var requestMessage = new HttpRequestMessage(HttpMethod.Post, "https://api-free.deepl.com/v2/translate")
-                {
-                    Content = new StringContent(JsonSerializer.Serialize(requestBody), Encoding.UTF8, DefaultContentType),
-                    Headers = { { HttpRequestHeader.Authorization.ToString(), $"DeepL-Auth-Key {Service.configuration.DeepL_API_Key}" } }
-                };
-
-                try
-                {
-                    var response = await HttpClient.SendAsync(requestMessage).ConfigureAwait(false);
-                    response.EnsureSuccessStatusCode();
-                    var jsonResponse = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                    var translated = JObject.Parse(jsonResponse)["translations"]?[0]?["text"]?.ToString().Trim();
-                    return !string.IsNullOrWhiteSpace(translated)
-                        ? (targetLanguage == "Chinese (Traditional)" ? await MachineTranslate(translated, "Chinese (Traditional)") : translated)
-                        : throw new Exception("Translation not found in the expected JSON structure.");
-                }
-                catch
-                {
-                    return await MachineTranslate(text, targetLanguage);
-                }
-            }
-            return "Target language not supported by DeepL.";
-        }
-
-        public static bool TryGetLanguageCode(string language, out string? languageCode)
-        {
-            languageCode = language switch
-            {
-                "English" => "EN-GB",
-                "Japanese" => "JA",
-                "German" => "DE",
-                "French" => "FR",
-                "Korean" => "KO",
-                "Chinese (Simplified)" => "ZH",
-                "Chinese (Traditional)" => "ZH",
-                "Spanish" => "ES",
-                _ => null
-            };
-            return !string.IsNullOrEmpty(languageCode);
-        }
-
-        public static async Task<string> LLMProxyTranslate(string message, string targetLanguage)
-        {
+                var language = await Translator.GTranslator.DetectLanguageAsync(messageText);
+                Service.pluginLog.Debug($"{messageText}\n -> language: {language.Name}");
 #if DEBUG
-            string Cfv2 = Service.configuration.Proxy_API_Key;
-#else
-            if (string.IsNullOrEmpty(Cfv2))
-            {
-                return await MachineTranslate(message, targetLanguage);
-            }
+                Plugin.OutputChatLine($"{messageText}\n -> language: {language.Name}");
 #endif
-
-            var requestData = new { regionCode = Service.configuration.ProxyRegion, targetLanguage, message };
-            var request = new HttpRequestMessage(HttpMethod.Post, "https://cfv2.kelpcc.com")
-            {
-                Content = new StringContent(JsonSerializer.Serialize(requestData), Encoding.UTF8, DefaultContentType)
-            };
-            request.Headers.Add("x-api-key", Cfv2);
-
-            try
-            {
-                var response = await HttpClient.SendAsync(request).ConfigureAwait(false);
-                response.EnsureSuccessStatusCode();
-                var responseBody = await response.Content.ReadAsStringAsync();
-                var translated = JObject.Parse(responseBody)["translated"]?.ToString().Trim();
-                return !string.IsNullOrWhiteSpace(translated) && translated != "{}" ? translated.Replace("\n", string.Empty) : throw new Exception("Translation not found in the expected JSON structure.");
+                langStr = language.Name;
             }
-            catch
+            catch (Exception GTex)
             {
-                return await MachineTranslate(message, targetLanguage);
-            }
-        }
-
-        private static async Task<string> OpenAITranslate(string message, string targetLanguage)
-        {
-            if (!Regex.IsMatch(Service.configuration.OpenAI_API_Key, @"^sk-[a-zA-Z0-9]{32,}$"))
-            {
-                return await MachineTranslate(message, targetLanguage);
-            }
-
-            var prompt = $"Translate the following FFXIV chat message into {targetLanguage}. If you encounter any in-game terms, keep them in their original form. Output only the translated text in a single line.\nMessage to translate: {message}";
-            var requestData = new
-            {
-                model = "gpt-3.5-turbo",
-                temperature = 0.6,
-                max_tokens = Math.Min(Math.Max(message.Length * 2, 20), 150),
-                messages = new[]
+                Service.pluginLog.Warning($"Google Translate failed to detect language. {GTex}");
+                try
                 {
-                    new { role = "system", content = prompt },
-                    new { role = "user", content = message }
+                    var language = await Translator.BingTranslator.DetectLanguageAsync(messageText);
+                    Service.pluginLog.Debug($"{messageText}\n -> language: {language.Name}");
+                    langStr = language.Name;
                 }
-            };
-
-            var request = new HttpRequestMessage(HttpMethod.Post, "https://api.openai.com/v1/chat/completions")
-            {
-                Content = new StringContent(JsonSerializer.Serialize(requestData), Encoding.UTF8, DefaultContentType),
-                Headers = { { HttpRequestHeader.Authorization.ToString(), $"Bearer {Service.configuration.OpenAI_API_Key}" } }
-            };
-
-            try
-            {
-                var response = await HttpClient.SendAsync(request).ConfigureAwait(false);
-                response.EnsureSuccessStatusCode();
-                var jsonResponse = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                var translated = JObject.Parse(jsonResponse)["choices"]?[0]?["message"]?["content"]?.ToString().Trim();
-                return !string.IsNullOrWhiteSpace(translated) ? translated : throw new Exception("Translation not found in the expected JSON structure.");
+                catch (Exception BTex)
+                {
+                    Service.pluginLog.Warning($"Bing Translate failed to detect language. {BTex}");
+                }
             }
-            catch
-            {
-                return await MachineTranslate(message, targetLanguage);
-            }
+            return langStr ?? "unknown";
         }
     }
 }
